write reference resolver with typewriter
