+++ go.mod
@@ -1,13 +1,14 @@
 module github.com/crossplane/crossplane-tools
 
-go 1.12
+go 1.16
 
 require (
 	github.com/alecthomas/template v0.0.0-20190718012654-fb15b899a751 // indirect
 	github.com/alecthomas/units v0.0.0-20190924025748-f65c72e2690d // indirect
 	github.com/dave/jennifer v1.3.0
-	github.com/google/go-cmp v0.3.1
-	github.com/pkg/errors v0.8.1
-	golang.org/x/tools v0.0.0-20191018212557-ed542cd5b28a
+	github.com/google/go-cmp v0.5.4
+	github.com/muvaf/typewriter v0.0.0-20210830125036-2204b298ddcf
+	github.com/pkg/errors v0.9.1
+	golang.org/x/tools v0.1.0
 	gopkg.in/alecthomas/kingpin.v2 v2.2.6
 )
